Exported ZLib to a separate static library
With this stage, to better relfect the architecture, I moved ZLib to a separate static library within the project.
